# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  message: string;
}

export interface AuthUser {
  id: string;
  name: string;
  email: string;
}

export interface AuthUserResponse {
  user: AuthUser;
}

export interface RegisterInput {
  /**
     * @minLength 2
     * @maxLength 40
     */
  name: string;
  email: string;
  /**
     * @minLength 8
     * @maxLength 128
     */
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export type UserDataInputData = { [key: string]: unknown };

export interface UserDataInput {
  data: UserDataInputData;
}

export type UserDataResponseData = { [key: string]: unknown };

export interface UserDataResponse {
  data: UserDataResponseData;
}

/**
 * Bad request
 */
export type BadRequestResponse = ErrorResponse;

/**
 * Conflict
 */
export type ConflictResponse = ErrorResponse;

/**
 * Authentication required
 */
export type UnauthorizedResponse = ErrorResponse;