# lib/api-zod/src/generated/types/userDataResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { UserDataResponseData } from './userDataResponseData';

export interface UserDataResponse {
  data: UserDataResponseData;
}